Moved images/frontend view module to images/serve
