+++ wagtail/wagtailimages/urls.py
@@ -2,8 +2,8 @@ from __future__ import absolute_import, unicode_literals
 
 from django.conf.urls import url
 
-from wagtail.wagtailimages.views import frontend
+from wagtail.wagtailimages.views.serve import serve
 
 urlpatterns = [
-    url(r'^(.*)/(\d*)/(.*)/$', frontend.serve, name='wagtailimages_serve'),
+    url(r'^(.*)/(\d*)/(.*)/$', serve, name='wagtailimages_serve'),
 ]
