+++ wagtail/wagtailimages/views/serve.py
